Define "go" and "singularity" targets in the BUILD_DIR
By moving the build to a subdirectory `BUILD_DIR`, make was not aware
anymore if the "go" and "singularity" targets already exists and thus
they were downloaded/rebuild every time.

To fix this, add the `${BUILD_DIR}/`-prefix to the target names.  Also
change the target name for the build directory to the actual directory
path.

Note that the `|` for the dependency of the "go" target is important as
it tells make to only check if `${BUILD_DIR}` exists and not consider
its last modification date.  Without this, "go" would still be rebuild
every time because the modification date of a directory is updated
every time something inside the directory is changed (which happens in
every build).