fix: VERSION needs to be in the root directory
Unfortunately the `VERSION` file has to be moved to the root directory
of this repository to be found, so inside the build directory is not
enough.  This is because the Singularity build script tries to find it
at the root of the git repository if git is detected.

The build was still successful without this fix, however, `singularity
--version` was then showing

    singularity version cfdb5d4-dirty

instead of the proper version.  Note that the hash in this case refers
to this repository, not something related to the Singularity source.